style: 更改文件名

+++ src/utils/utils_string.c
@@ -1,28 +1,27 @@
 #include <ctype.h>
 #include <stdio.h>
-#include <stdlib.h>
 
-#include "string_.h"
+#include "utils_string.h"
 
 
-int str2lower(char* str, int len)
+int str2lower(char* str, size_t len)
 {
     if (str == NULL || len < 0)
         return -1;
 
-    for (int i = 0; str[i] && i < len; i++)
+    for (size_t i = 0; str[i] && i < len; i++)
         str[i] = tolower(str[i]);
 
     return 0;
 }
 
 
-int str2upper(char* str, int len)
+int str2upper(char* str, size_t len)
 {
     if (str == NULL || len < 0)
         return -1;
 
-    for (int i = 0; str[i] && i < len; i++)
+    for (size_t i = 0; str[i] && i < len; i++)
         str[i] = toupper(str[i]);
 
     return 0;
